LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/bailout-side-effects-synchronous-script.window-expected.txt


PASS document.open bailout should not have any side effects (active parser whose script nesting level is greater than 0)
